requirements.txt
streamlit==1.35.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.1
sentence-transformers==2.2.2
numpy==1.24.3
chromadb==0.4.24
huggingface-hub==0.22.2